Fix 32bit compilation issue

+++ modules/packet-pcap/main.c
@@ -300,7 +300,7 @@ static int packet_do_resize(struct packet *orig_pkt, size_t size)
 static uint64 packet_get_id(struct packet *orig_pkt)
 {
 	struct pcap_packet *pkt = (struct pcap_packet*)orig_pkt;
-	return (pkt->header.ts.tv_sec << 32) + pkt->header.ts.tv_usec;
+	return (((uint64)pkt->header.ts.tv_sec) << 32) + pkt->header.ts.tv_usec;
 }
 
 static const char *packet_get_dissector(struct packet *pkt)
